requirements.txt
Django>=1.4
django-taggit>=0.9.3
django_evolution>=0.6.7